Moving Graph and ProxyGraph implementations into mgt
