+++ src/mgt-core/utils/GraphHelpers.ts
@@ -17,7 +17,6 @@ import {
 } from '@microsoft/microsoft-graph-client';
 import * as MicrosoftGraph from '@microsoft/microsoft-graph-types';
 import { PACKAGE_VERSION } from '../../version';
-import { Graph } from '../Graph';
 import { IGraph } from '../IGraph';
 import { IProvider } from '../providers/IProvider';
 import { SdkVersionMiddleware } from './SdkVersionMiddleware';
@@ -95,29 +94,4 @@ export function blobToBase64(blob: Blob): Promise<string> {
     };
     reader.readAsDataURL(blob);
   });
-}
-
-/**
- * create a new Graph instance using the specified provider.
- *
- * @static
- * @param {IProvider} provider
- * @returns {Graph}
- * @memberof Graph
- */
-export function createFromProvider(provider: IProvider, version?: string, component?: Element): IGraph {
-  const middleware: Middleware[] = [
-    new AuthenticationHandler(provider),
-    new RetryHandler(new RetryHandlerOptions()),
-    new TelemetryHandler(),
-    new SdkVersionMiddleware(PACKAGE_VERSION),
-    new HTTPMessageHandler()
-  ];
-
-  const client = Client.initWithMiddleware({
-    middleware: chainMiddleware(...middleware)
-  });
-
-  const graph = new Graph(client, version);
-  return component ? graph.forComponent(component) : graph;
 }
