+++ src/mgt/mock/MockGraph.ts
@@ -15,7 +15,8 @@ import {
   RetryHandlerOptions,
   TelemetryHandler
 } from '@microsoft/microsoft-graph-client';
-import { chainMiddleware, Graph, MgtBaseComponent } from '../../mgt-core';
+import { chainMiddleware, MgtBaseComponent } from '../../mgt-core';
+import { Graph } from '../Graph';
 import { MockProvider } from './MockProvider';
 
 /**
