+++ src/mgt/providers/SharePointProvider.ts
@@ -5,7 +5,8 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { createFromProvider, IProvider, ProviderState } from '../../mgt-core';
+import { IProvider, ProviderState } from '../../mgt-core';
+import { createFromProvider } from '../Graph';
 
 /**
  * AadTokenProvider
